Word Cloud: watch container resize and update visualization
