+++ client/app/pages/dashboards/dashboard.less
@@ -83,6 +83,7 @@
     .sunburst-visualization-container,
     .sankey-visualization-container,
     .map-visualization-container,
+    .word-cloud-visualization-container,
     .plotly-chart-container {
       position: absolute;
       left: 0;
