+++ client/app/visualizations/word-cloud/Renderer.jsx
@@ -2,6 +2,7 @@ import d3 from 'd3';
 import cloud from 'd3-cloud';
 import { each, map, min, max, values, sortBy } from 'lodash';
 import React, { useMemo, useState, useEffect } from 'react';
+import resizeObserver from '@/services/resizeObserver';
 import { RendererPropTypes } from '@/visualizations';
 
 import './renderer.less';
@@ -172,6 +173,13 @@ export default function Renderer({ data, options }) {
       render(container, words);
     }
   }, [container, words]);
+
+  useEffect(() => resizeObserver(container, () => {
+    const svg = container.querySelector('svg');
+    if (svg) {
+      scaleElement(svg, container);
+    }
+  }), [container]);
 
   return (<div className="word-cloud-visualization-container" ref={setContainer} />);
 }
