+++ client/app/visualizations/word-cloud/renderer.less
@@ -1,7 +1,12 @@
 .word-cloud-visualization-container {
   overflow: hidden;
+  height: 400px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
 
   svg {
-    transform-origin: left top;
+    transform-origin: center center;
+    flex: 0 0 auto;
   }
 }
